requirements.txt
jinja2>=3.1.6
pandas>=2.2.3
pyarrow>=19.0.1
pyyaml>=6.0.2
safetensors>=0.5.3
tensorboard>=2.19.0
tokenizers>=0.21.1
huggingface_hub
wandb